enhancement #35: migrate to pub.dev package format
