+++ analysis_options.yaml
@@ -1,4 +1,7 @@
+include: package:flutter_lints/flutter.yaml
 analyzer:
   errors:
     must_be_immutable: ignore
-include: package:flutter_lints/flutter.yaml
+
+# Additional information about this file can be found at
+# https://dart.dev/guides/language/analysis-options
