+++ example/android/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
