+++ example/lib/main.dart
@@ -1,9 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_math_fork/flutter_math.dart';
-import 'package:math_keyboard/keyboard_model.dart';
-import 'package:math_keyboard/widgets/keyboard.dart';
-import 'package:math_keyboard/widgets/math_input.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_math_input/flutter_math_input.dart';
 
 void main() {
   runApp(const MainApp());
