+++ lib/flutter_math_input.dart
@@ -1,0 +1,6 @@
+export 'keyboard_controller.dart';
+export 'widgets/keyboard/keyboard.dart';
+export 'widgets/keyboard/basic_keyboard/basic_keyboard.dart';
+export 'widgets/math_input.dart'; 
+export 'package:provider/provider.dart';
+export 'widgets/keyboard/custom_math_icons_icons.dart';
