+++ lib/fonts/CustomMathIcons.ttf
