+++ lib/parsers/formula_to_tex_parser.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:math_keyboard/services/math_constructions_building.dart';
-import 'package:math_keyboard/services/text_field_handle_and_create.dart';
+
+import '../services/math_constructions_building.dart';
+import '../services/text_field_handle_and_create.dart';
 
 class FormulaToTexParser {
   String formulaInTeX = '';
