+++ lib/services/math_constructions_building.dart
@@ -1,7 +1,8 @@
 // ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:math_keyboard/services/text_field_handle_and_create.dart';
+
+import 'text_field_handle_and_create.dart';
 
 enum ElementsType {
   fracElement,
