+++ lib/services/widgets_data_handler.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:math_keyboard/parsers/formulas_tree_parsers.dart';
-import 'package:math_keyboard/services/math_constructions_building.dart';
+import '../parsers/formulas_tree_parsers.dart';
+import 'math_constructions_building.dart';
+
 
 class WidgetsDataHandler{
   void replaceWidgetInTree(ParsedWidgetsData parsedData, MathConstructionData mathConstructions){
@@ -12,7 +13,7 @@ class WidgetsDataHandler{
   }
   void addToWidgetTree(ParsedWidgetsData parsedData, List<Widget> widget){
     parsedData.wigetData!.addAll(widget); 
-  }
+  } 
   void deleteFromWidgetTree(ParsedWidgetsData parsedData){
     parsedData.wigetData?.removeAt(parsedData.index!); 
   }
