+++ lib/widgets/keyboard/basic_keyboard/bacis_keyboard_pages/standart_numbers_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:math_keyboard/custom_math_icons_icons.dart';
-import 'package:math_keyboard/keyboard_model.dart';
-import 'package:math_keyboard/widgets/float_button_overlay.dart';
-import 'package:math_keyboard/widgets/keyboard.dart';
+import 'package:flutter_math_input/keyboard_controller.dart';
+import 'package:flutter_math_input/widgets/keyboard/basic_keyboard/float_button_overlay.dart';
+import '../basic_keyboard.dart';
+import '../../custom_math_icons_icons.dart';
 
 class NumbersPageWidget extends StatelessWidget {
   const NumbersPageWidget({
