+++ lib/widgets/keyboard/custom_math_icons_icons.dart
@@ -1,18 +1,3 @@
-/// Flutter icons CustomMathIcons
-/// Copyright (C) 2024 by original authors @ fluttericon.com, fontello.com
-/// This font was generated by FlutterIcon.com, which is derived from Fontello.
-///
-/// To use this font, place it in your fonts/ directory and include the
-/// following in your pubspec.yaml
-///
-/// flutter:
-///   fonts:
-///    - family:  CustomMathIcons
-///      fonts:
-///       - asset: fonts/CustomMathIcons.ttf
-///
-/// 
-///
 // ignore_for_file: constant_identifier_names
 
 import 'package:flutter/widgets.dart';
