+++ pubspec.yaml
@@ -1,28 +1,59 @@
-name: math_keyboard
-description: "A new Flutter project."
-publish_to: 'none'
-version: 0.1.0
+name: flutter_math_input
+description: "Package for displaying, inputting, and outputting mathematical constructs in KaTeX."
+version: 0.0.1
+homepage:
 
 environment:
-  sdk: '>=3.4.4 <4.0.0'
+  sdk: ^3.6.1
+  flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
   provider: ^6.1.2
-  flutter_math_fork: ^0.7.3
 
-
-  
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^5.0.0
   test: ^1.25.8
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
 
+# The following section is specific to Flutter packages.
 flutter:
   uses-material-design: true
   fonts:
     - family:  CustomMathIcons
       fonts:
-       - asset: fonts/CustomMathIcons.ttf
+       - asset: lib/fonts/CustomMathIcons.ttf
+  # To add assets to your package, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+  #
+  # For details regarding assets in packages, see
+  # https://flutter.dev/to/asset-from-package
+  #
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/to/resolution-aware-images
+
+  # To add custom fonts to your package, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts in packages, see
+  # https://flutter.dev/to/font-from-package
