+++ test/text_field_service_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_math_input/services/text_field_handle_and_create.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:math_keyboard/services/text_field_handle_and_create.dart';
 
 void main() {
   group('Focus tests', () {
